refactor (notification-service): refactored EmailStrategyFactory and EmailService
